requirements.txt
pip
fastapi
uvicorn
langchain-google-genai
pydantic
langchain
python-dotenv